4422: Move ConsoleOutput and PropertiesDiff files back to client module
