+++ java/clients/src/main/java/sleeper/clients/admin/properties/PropertyDiff.java
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-package sleeper.core.properties;
+package sleeper.clients.admin.properties;
 
-import sleeper.core.util.ConsoleOutput;
+import sleeper.clients.util.console.ConsoleOutput;
+import sleeper.core.properties.SleeperProperty;
+import sleeper.core.properties.SleeperPropertyIndex;
 
 import java.util.Map;
 import java.util.Objects;
