+++ java/clients/src/main/java/sleeper/clients/util/console/ConsoleOutput.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.core.util;
+package sleeper.clients.util.console;
 
 import java.io.PrintStream;
 import java.io.PrintWriter;
