+++ java/clients/src/test/java/sleeper/clients/admin/CompactionStatusReportScreenTest.java
@@ -50,8 +50,8 @@ import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.TASK_QU
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.TASK_QUERY_UNFINISHED_OPTION;
 import static sleeper.clients.status.report.compaction.task.CompactionTaskStatusReportTestHelper.startedTask;
 import static sleeper.clients.testutil.TestConsoleInput.CONFIRM_PROMPT;
+import static sleeper.clients.util.console.ConsoleOutput.CLEAR_CONSOLE;
 import static sleeper.core.properties.instance.CompactionProperty.COMPACTION_TRACKER_ENABLED;
-import static sleeper.core.util.ConsoleOutput.CLEAR_CONSOLE;
 
 class CompactionStatusReportScreenTest extends AdminClientMockStoreBase {
     @Nested
