+++ java/clients/src/test/java/sleeper/clients/admin/TableNamesReportTest.java
@@ -30,8 +30,8 @@ import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.MAIN_SC
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.PROMPT_RETURN_TO_MAIN;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.TABLE_NAMES_REPORT_OPTION;
 import static sleeper.clients.testutil.TestConsoleInput.CONFIRM_PROMPT;
+import static sleeper.clients.util.console.ConsoleOutput.CLEAR_CONSOLE;
 import static sleeper.core.table.TableStatus.uniqueIdAndName;
-import static sleeper.core.util.ConsoleOutput.CLEAR_CONSOLE;
 
 class TableNamesReportTest extends AdminClientMockStoreBase {
