+++ java/clients/src/test/java/sleeper/clients/testutil/TestConsoleInput.java
@@ -16,7 +16,7 @@
 package sleeper.clients.testutil;
 
 import sleeper.clients.util.console.ConsoleInput;
-import sleeper.core.util.ConsoleOutput;
+import sleeper.clients.util.console.ConsoleOutput;
 
 import java.util.Arrays;
 import java.util.LinkedList;
